# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Financemanagment API Documentation",
        "description": "This is auto-generated Swagger API documentation.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/goals": {
            "get": {
                "tags": [
                    "Goals"
                ],
                "summary": "List all goals",
                "operationId": "88058068651db3e85fb1d15aedba880c",
                "responses": {
                    "200": {
                        "description": "List of goals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GoalResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Goals"
                ],
                "summary": "Create a new goal",
                "operationId": "f398c4e52ba126f80633b34d5cc23690",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GoalCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Goal created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoalResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/goals/{goal}": {
            "get": {
                "tags": [
                    "Goals"
                ],
                "summary": "Get a single goal",
                "operationId": "86bd9e91921ff24f1cca03016f9f83df",
                "parameters": [
                    {
                        "name": "goal",
                        "in": "path",
                        "description": "ID of goal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoalResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Goal not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Goals"
                ],
                "summary": "Update a goal",
                "operationId": "8284bcf9d28f8c08d4beda6869d5cb78",
                "parameters": [
                    {
                        "name": "goal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GoalUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated goal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoalResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Goals"
                ],
                "summary": "Delete a goal",
                "operationId": "0f7f125ae930572f80815853318ea875",
                "parameters": [
                    {
                        "name": "goal",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Goal deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Goal deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Goal not found"
                    }
                }
            }
        },
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions",
                "operationId": "9b50e29e364179ebde24ac5fc59cd442",
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transaction"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "operationId": "54eaa3952808f639d46f969f4c8b98c5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/{transaction}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get a single transaction",
                "operationId": "846b687ea89d64602c002df14e07b802",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update a transaction",
                "operationId": "df36b362db8d9d4b25b5c8a24053a8a2",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated transaction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Delete a transaction",
                "operationId": "cdb3a161edf4b91189ecc95d452d98eb",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Category",
                "required": [
                    "id",
                    "user_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Uyga"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "GoalResource": {
                "title": "Goal",
                "required": [
                    "id",
                    "user_id",
                    "title",
                    "target_amount"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "title": {
                        "type": "string",
                        "example": "Umra uchun tejash!"
                    },
                    "target_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 15000000
                    },
                    "current_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 250000
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "GoalCreate": {
                "required": [
                    "title",
                    "target_amount",
                    "deadline"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Umra uchun tejash!"
                    },
                    "target_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 15000000
                    },
                    "current_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "GoalUpdate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Umra uchun tejash!"
                    },
                    "target_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 15000000
                    },
                    "current_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 250000
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Transaction": {
                "required": [
                    "id",
                    "user_id",
                    "amount",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "type": {
                        "type": "string",
                        "example": "income"
                    },
                    "note": {
                        "type": "string",
                        "example": "Some note"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-06"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "entity": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "TransactionCreate": {
                "required": [
                    "amount",
                    "type"
                ],
                "properties": {
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "type": {
                        "type": "string",
                        "example": "income"
                    },
                    "note": {
                        "type": "string",
                        "example": "Some note"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-06"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "entity_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\Goal"
                    },
                    "entity_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "TransactionUpdate": {
                "properties": {
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "type": {
                        "type": "string",
                        "example": "income"
                    },
                    "note": {
                        "type": "string",
                        "example": "Some note"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-06"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "entity_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\Goal"
                    },
                    "entity_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
...
}